whisper-service/requirements-document.txt
flask==2.3.3
pypdf2==3.0.1
python-docx==0.8.11
python-pptx==0.6.21
pandoc==2.3
requests==2.31.0
